requirements.txt
Flask-RESTful
Flask-SQLAlchemy
flasgger